move rect + circle over to grid_shape

+++ crates/common/src/geometry/math/scale_points.rs
@@ -1,4 +1,5 @@
 use crate::prelude::*;
+
 /// Scale `points` (move them towards or away) around the `center` by `factor`
 ///
 /// The resulting points distance will be points[x].distance(center) * factor but at the
@@ -7,7 +8,7 @@ pub fn scale_points(center: impl Point2d, points: &[IVec2], factor: f32) -> Vec<
     let mut output = vec![];
     for point in points {
         let angle = center.angle_to(*point);
-        let dist = DistanceAlg::Pythagoras.distance2d(center, *point) as f32 * factor;
+        let dist = DistanceAlg::Pythagoras.distance2d(center, *point) * factor;
         output.push(IVec2::from_angle(center, dist.max(0.0), angle));
     }
     output
